+++ ensure-no-std/src/compat_test.rs
@@ -75,10 +75,10 @@ impl_more::impl_error_enum!(Errors: Wrapped(err) => err);
 impl_more::impl_enum_from!(LeafErr => Errors::Wrapped);
 
 enum GenericError<T> {
-    Other(T),
+    Other { source: T },
 }
 
-impl_more::impl_enum_from!(<T> in T => GenericError<T>::Other);
+impl_more::impl_enum_from!(<T> in T => GenericError<T>::Other { source });
 
 #[derive(Debug, Clone)]
 struct Checked(bool);