+++ crates/impl-more/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Add `impl_enum_from!` macro for conversions into enum variants with one unnamed field, including generic enums.
+- Add `impl_enum_from!` macro for conversions into enum variants with one named or unnamed field, including generic enums.
 
 ## 0.3.7